Refresh facts by reparsing jboss configuration file
